refactor(native-agent): extract the loopback ingress
Move the per-server loopback listener (ephemeral port, per-listener token,
/<scope>/<sessionId> posts) out of HookObservedAgentServer into
LoopbackIngress so other backends can report over the same channel.
No behavior change.

Agent-Session: ba208f4e-0500-4cca-bc8c-67aafcc19811
Agent-Session: 01a0600c-8adb-7663-ba10-a0207d0507b8
Agent-Session: 01a0602c-1d0f-7b03-8356-7a3614e78dc3
Agent-Session: 01a06048-aee8-7142-84b8-ef4838717f58
Agent-Session: 01a06050-6825-7661-94cc-512afb8a405e
Agent-Session: 01a060a7-1371-7d61-8a00-a54bee4571bd
Agent-Session: 3cb7af5b-fbd3-47a0-b1fb-500205008caf
Agent-Session: bf4b4f2e-5a25-4a22-a6ae-a7a3e65512f8
Agent-Session: ses_f9870de22ffeji2kebmLyh6jOy
