+++ packages/progressive-review/src/native-agent/hook-observed-server.ts
@@ -1,10 +1,7 @@
-import { randomBytes } from "node:crypto";
-import { type IncomingMessage, type Server, createServer } from "node:http";
-import type { AddressInfo } from "node:net";
-
 import { DEV_REVIEW_HOME_ENV, devReviewHome } from "../review-storage";
 import { AsyncQueue } from "./async-queue";
 import type { HarnessDialect, NativeTerminalCommand } from "./harness";
+import { LoopbackIngress } from "./loopback-ingress";
 import type {
   AgentServer,
   AgentServerOptions,
@@ -23,8 +20,6 @@ import {
 } from "./terminal-command";
 import { isJsonRecord } from "./transcript-json";
 
-const MAX_HOOK_BODY_BYTES = 2 * 1024 * 1024;
-
 interface SessionState {
   transcriptPath?: string;
   subscribers: Set<Subscriber>;
@@ -40,11 +35,11 @@ interface Subscriber {
 }
 
 /**
- * An AgentServer for harnesses that have no server of their own. It listens
- * on a loopback port; the terminal's native hooks post there, and each hook
- * is a signal to re-read the transcript and forward whatever is new. Nothing
- * here is harness-specific: the dialect supplies the command line and the
- * reader.
+ * An AgentServer for harnesses that have no server of their own. The
+ * terminal's native hooks post to this server's loopback ingress, and each
+ * hook is a signal to re-read the transcript and forward whatever is new.
+ * Nothing here is harness-specific: the dialect supplies the command line
+ * and the reader.
  */
 export class HookObservedAgentServer implements AgentServer {
   readonly harness: HarnessDialect["harness"];
@@ -53,8 +48,7 @@ export class HookObservedAgentServer implements AgentServer {
   readonly #desktop: AgentServerOptions["desktopEndpoint"];
   readonly #commandPath: ReviewCommandPath;
   readonly #sessions = new Map<string, SessionState>();
-  readonly #hookToken = randomBytes(32).toString("base64url");
-  #listener: Promise<{ server: Server; url: string }> | undefined;
+  readonly #ingress: LoopbackIngress;
 
   constructor(dialect: HarnessDialect, options: AgentServerOptions) {
     this.harness = dialect.harness;
@@ -65,6 +59,11 @@ export class HookObservedAgentServer implements AgentServer {
       token: options.desktopEndpoint.token,
     };
     this.#commandPath = new ReviewCommandPath(options);
+    this.#ingress = new LoopbackIngress({
+      scope: this.harness,
+      onPost: (sessionId, payload) =>
+        this.#receiveHookEvent(sessionId, payload),
+    });
   }
 
   async launch(
@@ -74,7 +73,7 @@ export class HookObservedAgentServer implements AgentServer {
       session: input.session,
       cwd: input.cwd,
     });
-    const { url: hookBaseUrl } = await this.#listen();
+    const hookBaseUrl = await this.#ingress.url();
     const sessionPath = `${this.harness}/${encodeURIComponent(sessionId)}`;
     const pathValue = await this.#commandPath.resolve();
     const command = await this.#dialect.terminalCommand({
@@ -85,7 +84,7 @@ export class HookObservedAgentServer implements AgentServer {
       env: {
         // Native hooks report to this server's own listener.
         [REVIEW_AGENT_HOOK_URL_ENV]: `${hookBaseUrl}/${sessionPath}`,
-        [REVIEW_AGENT_HOOK_TOKEN_ENV]: this.#hookToken,
+        [REVIEW_AGENT_HOOK_TOKEN_ENV]: this.#ingress.token,
         // `review threads get` inside the terminal reads its thread from the
         // desktop, which every implementation needs regardless of observation.
         [REVIEW_AGENT_THREAD_URL_ENV]: `${this.#desktop.baseUrl}/native-agent-events/${sessionPath}/thread`,
@@ -132,66 +131,7 @@ export class HookObservedAgentServer implements AgentServer {
       state.subscribers.clear();
     }
     await Promise.all(pending);
-    if (this.#listener) {
-      const { server } = await this.#listener;
-      this.#listener = undefined;
-      server.closeAllConnections();
-      await new Promise<void>((resolve, reject) => {
-        server.close((error) => (error ? reject(error) : resolve()));
-      });
-    }
-  }
-
-  /** Loopback listener for native hooks. Started on the first launch. */
-  #listen(): Promise<{ server: Server; url: string }> {
-    this.#listener ??= new Promise((resolve, reject) => {
-      const server = createServer((request, response) => {
-        void this.#handleHook(request)
-          .then(({ status, body }) => {
-            response.writeHead(status, { "content-type": "application/json" });
-            response.end(JSON.stringify(body));
-          })
-          .catch((error: unknown) => {
-            response.writeHead(500, { "content-type": "application/json" });
-            response.end(JSON.stringify({ ok: false, error: String(error) }));
-          });
-      });
-      server.once("error", reject);
-      server.listen(0, "127.0.0.1", () => {
-        const { port } = server.address() as AddressInfo;
-        resolve({ server, url: `http://127.0.0.1:${port}` });
-      });
-    });
-    return this.#listener;
-  }
-
-  async #handleHook(
-    request: IncomingMessage,
-  ): Promise<{ status: number; body: unknown }> {
-    if (request.method !== "POST") {
-      return { status: 405, body: { ok: false, error: "POST only." } };
-    }
-    if (request.headers["x-review-token"] !== this.#hookToken) {
-      return { status: 401, body: { ok: false, error: "Unauthorized" } };
-    }
-    const match = /^\/([^/]+)\/([^/]+)$/u.exec(request.url ?? "");
-    if (!match || match[1] !== this.harness) {
-      return { status: 404, body: { ok: false, error: "Unknown session." } };
-    }
-    const sessionId = decodeURIComponent(match[2]!);
-    const payload = await readJsonBody(request);
-    try {
-      this.#receiveHookEvent(sessionId, payload);
-    } catch (error) {
-      return {
-        status: 400,
-        body: {
-          ok: false,
-          error: error instanceof Error ? error.message : String(error),
-        },
-      };
-    }
-    return { status: 200, body: { ok: true } };
+    await this.#ingress.close();
   }
 
   #receiveHookEvent(sessionId: string, payload: unknown): void {
@@ -254,21 +194,6 @@ export class HookObservedAgentServer implements AgentServer {
       throw error;
     }
   }
-}
-
-async function readJsonBody(request: IncomingMessage): Promise<unknown> {
-  const chunks: Buffer[] = [];
-  let size = 0;
-  for await (const chunk of request) {
-    const buffer = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk);
-    size += buffer.length;
-    if (size > MAX_HOOK_BODY_BYTES) {
-      throw new Error("Native hook payload is too large.");
-    }
-    chunks.push(buffer);
-  }
-  const text = Buffer.concat(chunks).toString("utf8");
-  return text ? (JSON.parse(text) as unknown) : {};
 }
 
 function isMissingTranscript(error: unknown): boolean {
